BackendFaceRecognition/requirements.txt
numpy~=1.24.4
fastapi~=0.110.2
uvicorn~=0.29.0
pymongo~=4.6.3
python-multipart~=0.0.9
face-recognition~=1.3.0
pillow~=10.3.0
